fix(storage-plugin): restore inline value editing and extract entry detail dialog
Wraps handleValueChange/updateEntriesForSelectedStorage in useCallback so
the memoised columns array captures fresh client/selectedStorage state
instead of the stale first-render closure that silently no-oped every
inline value commit. Scopes stopPropagation in the value column to the
editable cell and edit button only, so clicking non-editable values still
opens the entry detail dialog. Moves EntryDetailDialog back out of
panel.tsx into entry-detail-dialog.tsx (with jsonSafeParse), and extracts
the shared formatValue JSX helper into format-value.tsx so panel.tsx and
the dialog don't duplicate it.
